// ui-kit/packages/styles/types/itheme.d.ts
import IPalette from './ipalette';

interface ITheme {
  baseFontSize: number;
  fontFamily: string;
  mainBlackColor: string;
  mainGrayColor: string;
  mainWhiteColor: string;
  palette: IPalette;
  mainBackgroundColor: string;
  h1FontSize: number;
  h2FontSize: number;
  h3FontSize: number;
  h4FontSize: number;
  h5FontSize: number;
  h6FontSize: number;
  mainOutlinedColor: string;
  mainOutlinedHoverColor: string;
  inactiveBackgroundColor: string;
  inactiveColor: string;
}

export default ITheme;